Lab 2: Traffic lights
Add util code to set mode of specific pins on GPIO ports.
Add util code to set pull up mode of specific pins on GPIO ports.

Add util code to get the status of a pulled up input on a GPIO port.
Add util code to change the status of an output port on a GPIO port.

Add code to make button press go thru sequence of:
- Green (pause)
- Yellow (pause)
- Red (pause)
- Red/Yellow (pause)
- Green

With debouncing.